Changes:
PhysicalScene now grabs its camera from sceneStack.Game.Services.
Refactored ItemFactory creation logic.
Window.Title is updated to camera and SceneStack depth info.;
BounceGame now pushes and pops new scenes exclusively (for now), but not if scene == typeof(BottomScene).
Paddle is now a Kinematic body. It moves when I set the velocity by pressing arrow keys, and uses MathHelper.SmoothStep() to slow smoothly to a stop. It is working great, but needs minor adjustment.
PaddleBall now fixes itself to or launches itself from its paddle when the spacebar is pressed, depending on which action is required. The Weld type joint was acting funny, so I jointed it to the world with a FixedPrismaticJoint on the X axis. I also jointed it to its paddle with a DistanceJoint. As you move the paddle, those two joints restrict the ball's movement to an oscillation on the x axis above the paddle, which slows to a stop over some time determined by joint properties.